move assembly loading into respective benchmark files
