+++ brainscore/benchmarks/README.md
@@ -1,3 +1,7 @@
+
+**Versioning**: Each benchmark is numbered by a version number. 
+                Whenever the benchmark produces different scores, the version number is increased.
 # Brain-Score Benchmarks
 
-**File structure**: one file per assembly with all the benchmarks inside.
+**File structure**: One file per assembly with all the benchmarks inside.
+                    This file contains the loading of the assembly (with eventual processing) and the benchmarks.
