+++ brainscore/benchmarks/majaj2015.py
@@ -1,19 +1,20 @@
-from brainscore.assemblies.private import assembly_loaders
-from brainscore.benchmarks._neural_common import build_benchmark, TimeFilteredAssemblyLoader
+import brainscore
+from brainscore.benchmarks._neural_common import NeuralBenchmark, average_repetition
 from brainscore.metrics.ceiling import InternalConsistency, RDMConsistency, TemporalCeiling
 from brainscore.metrics.rdm import RDMCrossValidated
 from brainscore.metrics.regression import CrossRegressedCorrelation, mask_regression, ScaledCrossRegressedCorrelation, \
     pls_regression, pearsonr_correlation
 from brainscore.metrics.temporal import TemporalRegressionAcrossTime, TemporalCorrelationAcrossImages
+from brainscore.utils import LazyLoad
 
 
 def _DicarloMajaj2015Region(region, identifier_metric_suffix, similarity_metric, ceiler):
-    return build_benchmark(f'dicarlo.Majaj2015.{region}-{identifier_metric_suffix}',
-                           assembly_loader=assembly_loaders[f'dicarlo.Majaj2015.private.{region}'],
-                           similarity_metric=similarity_metric,
-                           ceiler=ceiler,
-                           parent=region,
-                           paper_link='http://www.jneurosci.org/content/35/39/13402.short')
+    assembly_repetition = LazyLoad(lambda region=region: load_assembly(region))
+    assembly = LazyLoad(lambda: average_repetition(assembly_repetition))
+    return NeuralBenchmark(identifier=f'dicarlo.Majaj2015.{region}-{identifier_metric_suffix}',
+                           assembly=assembly, similarity_metric=similarity_metric,
+                           ceiling_func=lambda: ceiler(assembly_repetition),
+                           parent=region, paper_link='http://www.jneurosci.org/content/35/39/13402.short')
 
 
 def DicarloMajaj2015V4PLS():
@@ -62,16 +63,39 @@ def DicarloMajaj2015ITRDM():
                                    ceiler=RDMConsistency())
 
 
+def load_assembly(region):
+    assembly = brainscore.get_assembly(name='dicarlo.Majaj2015.private')
+    assembly = assembly.sel(region=region)
+    assembly['region'] = 'neuroid', [region] * len(assembly['neuroid'])
+    assembly = assembly.squeeze("time_bin")
+    assembly.load()
+    assembly = assembly.transpose('presentation', 'neuroid')
+    return assembly
+
+
 def _DicarloMajaj2015TemporalRegion(region):
     metric = CrossRegressedCorrelation(regression=TemporalRegressionAcrossTime(regression=pls_regression()),
                                        correlation=TemporalCorrelationAcrossImages(correlation=pearsonr_correlation()))
     # sub-select time-bins, and get rid of overlapping time bins
-    time_bins = [(time_bin_start, time_bin_start + 20) for time_bin_start in range(0, 231, 20)]
-    loader = TimeFilteredAssemblyLoader(assembly_loaders[f'dicarlo.Majaj2015.temporal.private.{region}'], time_bins)
-    return build_benchmark(identifier=f'dicarlo.Majaj2015.temporal.{region}', assembly_loader=loader,
-                           similarity_metric=metric, ceiler=TemporalCeiling(InternalConsistency()),
+    time_bins = tuple((time_bin_start, time_bin_start + 20) for time_bin_start in range(0, 231, 20))
+    assembly_repetition = LazyLoad(lambda region=region, time_bins=time_bins: load_temporal_assembly(region, time_bins))
+    assembly = LazyLoad(lambda: average_repetition((assembly_repetition)))
+    ceiler = TemporalCeiling(InternalConsistency())
+    return NeuralBenchmark(identifier=f'dicarlo.Majaj2015.temporal.{region}-pls_across_time',
+                           assembly=assembly, similarity_metric=metric,
+                           ceiling_func=lambda: ceiler(assembly_repetition),
                            parent=region, paper_link='http://www.jneurosci.org/content/35/39/13402.short')
 
 
 DicarloMajaj2015TemporalV4PLS = lambda: _DicarloMajaj2015TemporalRegion(region='V4')
 DicarloMajaj2015TemporalITPLS = lambda: _DicarloMajaj2015TemporalRegion(region='IT')
+
+
+def load_temporal_assembly(region, time_bins):
+    assembly = brainscore.get_assembly(name='dicarlo.Majaj2015.temporal.private')
+    assembly = assembly.sel(region=region)
+    assembly['region'] = 'neuroid', [region] * len(assembly['neuroid'])
+    assembly = assembly.sel(time_bin=time_bins)
+    assembly.load()
+    assembly = assembly.transpose('presentation', 'neuroid', 'time_bin')
+    return assembly
