+++ brainscore/metrics/ceiling.py
@@ -2,7 +2,7 @@ import scipy.stats
 import xarray as xr
 from tqdm import tqdm
 
-from brainscore.assemblies import walk_coords
+from brainio_base.assemblies import walk_coords
 from brainscore.metrics import Score
 from brainscore.metrics.rdm import RDMMetric
 from brainscore.metrics.transformations import CrossValidationSingle
