+++ brainscore/metrics/utils.py
@@ -2,7 +2,7 @@ from collections import OrderedDict
 
 import numpy as np
 
-from brainscore.assemblies import walk_coords
+from brainio_base.assemblies import walk_coords
 
 
 def collect_coords(assembly, ignore_dims, rename_coords_list, kind):
